10-deploy-ai-agent-langserve/requirements.txt
asana==5.0.7
python-dotenv==0.13.0
langchain==0.2.12
langserve==0.2.2
uvicorn==0.30.6
fastapi==0.112.2
gunicorn==23.0.0
sse_starlette==2.1.3
langchain-anthropic==0.1.22
langchain-groq==0.1.5
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-chroma==0.1.2
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
unstructured[all-docs]
streamlit==1.36.0
langgraph==0.1.19
aiosqlite==0.20.0
google-api-python-client==2.142.0
google-auth-oauthlib==1.2.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
httpx<0.26
copilotkit
